+++ core/core-api/src/services/proton.rs
@@ -79,10 +79,10 @@ use crate::{
 use bytes::Bytes;
 use parking_lot::RwLock;
 use proton_crypto_account::keys::APIPublicAddressKeys;
-use requests::DeleteContacts;
+use requests::PutDeleteContacts;
 use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
 use reqwest::{Client, Method, Url};
-use responses::{DeleteContactsResponse, GetAddressResponse};
+use responses::{GetAddressResponse, PutDeleteContactsResponse};
 use secrecy::ExposeSecret;
 use serde::{Deserialize, Serialize};
 use serde_json::{Error as JsonError, Value as JsonValue};
@@ -644,17 +644,17 @@ impl Proton {
             .await
     }
 
-    /// TODO: Document this method.
+    /// Method requests to delete contacts which remotes ids were provided.
     ///
     /// # Errors
     ///
     /// This method will return an error if the request fails.
     ///
-    pub async fn delete_contacts(
+    pub async fn put_delete_contacts(
         &self,
         ids: Vec<RemoteId>,
-    ) -> Result<DeleteContactsResponse, ApiServiceError> {
-        let body = DeleteContacts { ids };
+    ) -> Result<PutDeleteContactsResponse, ApiServiceError> {
+        let body = PutDeleteContacts { ids };
 
         self.put::<_, Json<_>>("contacts/delete", body, None).await
     }